lc solver - cleanup of steady state and time stepping lc solvers
