+++ qlc3d/includes/lc/lc-solver.h
@@ -27,6 +27,18 @@ struct LCSolverResult {
   const bool converged;
 };
 
+struct LCSolverParams {
+  double A;
+  double B;
+  double C;
+  double L1;
+  double L2;
+  double L3;
+  double L6;
+  double deleps;
+  double dt;
+  double u1;
+};
 
 class ILCSolver {
 public:
@@ -44,6 +56,10 @@ protected:
 
   /** Jacobian matrix */
   std::unique_ptr<SpaMtrix::IRCMatrix> K;
+  /** Euler-Lagrange equations vector */
+  std::unique_ptr<SpaMtrix::Vector> L;
+  /** The unknown vector of solutions for the LC */
+  std::unique_ptr<SpaMtrix::Vector> X;
 
   /**
    * Solve the matrix system Kx = r. The result is written into the vector x.
@@ -51,30 +67,14 @@ protected:
    */
   bool solveMatrixSystem(const SpaMtrix::IRCMatrix &Kmatrix, const SpaMtrix::Vector &r, SpaMtrix::Vector &x) const;
   double maxAbs(const SpaMtrix::Vector &v) const;
-public:
-  ImplicitLCSolver(const LC &lc, const SolverSettings &solverSettings);
 
-};
-
-
-
-class LCSolver : public ILCSolver, protected ImplicitLCSolver {
-protected:
-  const double rt2 = std::sqrt(2.0);
-  const double rt3 = std::sqrt(3.0);
-  const double rt6 = std::sqrt(6.0);
-  const double A, B, C;
-  const double L1, L2, L3, L6;
-  const double deleps;
-
-  std::unique_ptr<SpaMtrix::Vector> L;
-  std::unique_ptr<SpaMtrix::Vector> X;
-
-  void initialiseMatrixSystem(const SolutionVector &q, const Geometry &geom, double dt);
-  void assembleMatrixSystem(const SolutionVector &q, const SolutionVector &v, const Geometry &geom);
-  void assembleVolumeTerms(const SolutionVector &q, const SolutionVector &v, const Geometry &geom);
-  double updateQ(SolutionVector &q);
-
+  /**
+   * This is a common local element matrix assembly used by both steady-state and time-stepping solvers. This assembles
+   * the Jacobian matrix and the RHS vector (Euler-Lagrange equations) for a single tetrahedral element.
+   *
+   * In the time-stepping solver, the results of this method are further augmented by terms required by the
+   * implicit time-stepping algorithm.
+   */
   void assembleLocalVolumeMatrix(unsigned int indTet,
                                  double lK[20][20],
                                  double lL[20],
@@ -83,21 +83,53 @@ protected:
                                  GaussianQuadratureTet<11> shapes,
                                  const SolutionVector &q,
                                  const SolutionVector &v,
-                                 const Geometry &geom);
+                                 const Geometry &geom,
+                                 const LCSolverParams &params);
 
-  void addToGlobalMatrix(double lK[20][20],
-                         double lL[20],
-                         const SolutionVector &q,
-                         const unsigned int tetNodes[4],
-                         const unsigned int tetDofs[4]);
+  void addToGlobalMatrix(double lK[20][20], double lL[20], const SolutionVector &q, const unsigned int tetNodes[4]);
+  void assembleMatrixSystem(const SolutionVector &q, const SolutionVector &v, const Geometry &geom, const LCSolverParams &params);
+public:
+  ImplicitLCSolver(const LC &lc, const SolverSettings &solverSettings);
+};
+
+class SteadyStateLCSolver : public ILCSolver, protected ImplicitLCSolver {
+protected:
+  const double rt2 = std::sqrt(2.0);
+  const double rt3 = std::sqrt(3.0);
+  const double rt6 = std::sqrt(6.0);
+  const double A, B, C;
+  const double L1, L2, L3, L6;
+  const double deleps;
+
+  void initialiseMatrixSystem(const SolutionVector &q, const Geometry &geom, double dt);
+
 
 public:
-  virtual ~LCSolver();
-  LCSolver(const LC &lc, const SolverSettings &solverSettings);
+  virtual ~SteadyStateLCSolver();
+  SteadyStateLCSolver(const LC &lc, const SolverSettings &solverSettings);
 
   LCSolverResult solve(SolutionVector &q, const SolutionVector &v, const Geometry &geom, SimulationState &simulationState) override;
+};
 
+/**
+ * Time stepping LC solver implementing a non-linear implicit time stepping scheme where time derivatives are calculated
+ * at mid-points between time steps.
+ */
+class TimeSteppingLCSolver : public ILCSolver, protected ImplicitLCSolver {
+  /** Mass matrix, required by implicit time stepping */
+  std::unique_ptr<SpaMtrix::IRCMatrix> M;
+  /** Q-tensor at previous time step */
+  std::unique_ptr<SpaMtrix::Vector> q1;
+  /** RHS vector at previous time step */
+  std::unique_ptr<SpaMtrix::Vector> f_prev;
+  bool isFirstRun = true;
 
+  void initialiseMatrixSystem(const SolutionVector &q, const Geometry &geom);
+
+public:
+  TimeSteppingLCSolver(const LC &lc, const SolverSettings &solverSettings);
+  ~TimeSteppingLCSolver() = default;
+  LCSolverResult solve(SolutionVector &q, const SolutionVector &v, const Geometry &geom, SimulationState &simulationState) override;
 };
 
 #endif //PROJECT_QLC3D_LC_SOLVER_H
